[Refactoring] Move AddExplicitCodableImplementation to its own file
